+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - WMS-Time: Shrunk wms-time slider for mobile views, adjusted Position on desktop.
 - getFeatureInfo: gfi now works on BaseLayers as well as SubjectLayers.
 - Menutitle: Allow html in title text.
+- CompareMaps: layer selection fields are labeled left/right or upper/lower layer instead of first/second layer.
 
 ### Deprecated